GL2.0.0 COM_createHTMLDocument 使用 個別のデータでコメント停止の時でもコメントが表示される問題 データ新規作成時のコメントのデフォルトはコメント無効にする

+++ databox/public_html/data.php
@@ -187,8 +187,9 @@ function fncComment(
 	}
 	$mode = '';
 	if (isset ($_REQUEST['mode'])) {
-		$mode = COM_applyFilter ($_REQUEST['mode']);
+		$mode = COM_applyFilter ($_POST['mode']);
 	}
+	
 	$page = 1;
 	if (isset ($_REQUEST['cpage'])) {
 		$page = COM_applyFilter ($_REQUEST['cpage']);
@@ -226,21 +227,22 @@ function fncComment(
 	
     $result = DB_query ($sql);
     $numrows = DB_numRows ($result);
-
+	
     if ($numrows > 0) {
 
 		$A = DB_fetchArray ($result);
 		$A = array_map('stripslashes', $A);
-	
-		$delete_option = (SEC_hasRights('databox.edit') &&
+		if  ($A['commentcode']>=0){
+			$delete_option = (SEC_hasRights('databox.edit') &&
                     SEC_hasAccess($A['owner_id'], $A['group_id'],
                     $A['perm_owner'], $A['perm_group'], $A['perm_members'],
                     $A['perm_anon']) == 3 ? true : false);
 	
-		require_once $_CONF['path_system'] . 'lib-comment.php';
-		$retval .= CMT_userComments($id, $A['topic'], 'databox',
+			require_once $_CONF['path_system'] . 'lib-comment.php';
+			$retval .= CMT_userComments($id, $A['topic'], 'databox',
                                 $order, $mode, 0, $page, false,
                                 $delete_option, $A['commentcode']);
+		}
 	}
